refactor: Splitting column, index and foreign key comparison extension in different files
