+++ packages/serverpod/lib/src/database/extensions.dart
@@ -1,6 +1,9 @@
 import 'package:serverpod/protocol.dart';
 import 'package:serverpod/src/database/database_pool_manager.dart';
 import 'package:collection/collection.dart';
+import 'package:serverpod/src/database/extensions/column_comparison_extension.dart';
+import 'package:serverpod/src/database/extensions/foreign_key_comparison_extension.dart';
+import 'package:serverpod/src/database/extensions/index_comparison_extension.dart';
 import 'package:serverpod/src/database/migrations/table_comparison_warning.dart';
 
 /// Comparison methods for [DatabaseDefinition].
@@ -169,277 +172,6 @@ extension TableComparisons on TableDefinition {
   }
 }
 
-/// Comparison methods for [ColumnDefinition].
-extension ColumnComparisons on ColumnDefinition {
-  /// Compares this column definition with [other], returning a list of mismatches.
-  List<ColumnComparisonWarning> like(ColumnDefinition other) {
-    List<ColumnComparisonWarning> mismatches = [];
-
-    if (name != other.name) {
-      mismatches.add(
-        ColumnComparisonWarning(
-          name: name,
-          expected: name,
-          found: other.name,
-        ),
-      );
-    }
-
-    if (!columnType.like(other.columnType)) {
-      mismatches.add(
-        ColumnComparisonWarning(
-          name: 'type',
-          expected: '$columnType',
-          found: '${other.columnType}',
-        ),
-      );
-    }
-
-    if (isNullable != other.isNullable) {
-      mismatches.add(
-        ColumnComparisonWarning(
-          name: 'isNullable',
-          expected: '$isNullable',
-          found: '${other.isNullable}',
-        ),
-      );
-    }
-
-    if (columnDefault != other.columnDefault) {
-      mismatches.add(
-        ColumnComparisonWarning(
-          name: 'default value',
-          expected: '$columnDefault',
-          found: '${other.columnDefault}',
-        ),
-      );
-    }
-
-    return mismatches;
-  }
-}
-
-/// Comparison methods for [IndexDefinition].
-extension IndexComparisons on IndexDefinition {
-  /// Compares this index definition with [other], returning a list of mismatches.
-  List<IndexComparisonWarning> like(
-    IndexDefinition other, {
-    bool ignoreCase = false,
-  }) {
-    List<IndexComparisonWarning> mismatches = [];
-
-    if (ignoreCase &&
-            indexName.toLowerCase() != other.indexName.toLowerCase() ||
-        !ignoreCase && indexName != other.indexName) {
-      mismatches.add(
-        IndexComparisonWarning(
-          name: 'name',
-          expected: indexName,
-          found: other.indexName,
-        ),
-      );
-    }
-
-    if (type != other.type) {
-      mismatches.add(
-        IndexComparisonWarning(
-          name: 'type',
-          expected: type,
-          found: other.type,
-        ),
-      );
-    }
-
-    if (isUnique != other.isUnique) {
-      mismatches.add(
-        IndexComparisonWarning(
-          name: 'isUnique',
-          expected: '$isUnique',
-          found: '${other.isUnique}',
-        ),
-      );
-    }
-
-    if (isPrimary != other.isPrimary) {
-      mismatches.add(
-        IndexComparisonWarning(
-          name: 'isPrimary',
-          expected: '$isPrimary',
-          found: '${other.isPrimary}',
-        ),
-      );
-    }
-
-    if (predicate != other.predicate) {
-      mismatches.add(
-        IndexComparisonWarning(
-          name: 'predicate',
-          expected: predicate,
-          found: other.predicate,
-        ),
-      );
-    }
-
-    if (tableSpace != other.tableSpace) {
-      mismatches.add(
-        IndexComparisonWarning(
-          name: 'tableSpace',
-          expected: tableSpace,
-          found: other.tableSpace,
-        ),
-      );
-    }
-
-    if (elements.length != other.elements.length) {
-      mismatches.add(
-        IndexComparisonWarning(
-          name: 'elements',
-          expected: '${elements.length}',
-          found: '${other.elements.length}',
-        ),
-      );
-    } else {
-      for (int i = 0; i < elements.length; i++) {
-        var elementMismatches = elements[i].like(other.elements[i]);
-        if (elementMismatches.isNotEmpty) {
-          mismatches.addAll(elementMismatches);
-        }
-      }
-    }
-
-    return mismatches;
-  }
-}
-
-/// Comparison methods for [IndexElementDefinition].
-extension IndexElementDefinitionComparison on IndexElementDefinition {
-  /// Compares this index element with [other], returning a list of mismatches.
-  List<IndexComparisonWarning> like(IndexElementDefinition other) {
-    List<IndexComparisonWarning> mismatches = [];
-
-    if (type != other.type) {
-      mismatches.add(
-        IndexComparisonWarning(
-          name: 'element type',
-          expected: '$type',
-          found: '${other.type}',
-        ),
-      );
-    }
-
-    if (definition != other.definition) {
-      mismatches.add(
-        IndexComparisonWarning(
-          name: 'element definition',
-          expected: definition,
-          found: other.definition,
-        ),
-      );
-    }
-
-    return mismatches;
-  }
-}
-
-/// Comparison methods for [ForeignKeyDefinition].
-extension ForeignKeyComparisons on ForeignKeyDefinition {
-  /// Compares this foreign key definition with [other], returning a list of mismatches.
-  List<ForeignKeyComparisonWarning> like(
-    ForeignKeyDefinition other, {
-    bool ignoreCase = false,
-  }) {
-    List<ForeignKeyComparisonWarning> mismatches = [];
-
-    if (ignoreCase &&
-            constraintName.toLowerCase() !=
-                other.constraintName.toLowerCase() ||
-        !ignoreCase && constraintName != other.constraintName) {
-      mismatches.add(
-        ForeignKeyComparisonWarning(
-          name: 'name',
-          expected: constraintName,
-          found: other.constraintName,
-        ),
-      );
-    }
-
-    if (!const ListEquality().equals(columns, other.columns)) {
-      mismatches.add(
-        ForeignKeyComparisonWarning(
-          name: 'columns',
-          expected: '$columns',
-          found: '${other.columns}',
-        ),
-      );
-    }
-
-    if (referenceTable != other.referenceTable) {
-      mismatches.add(
-        ForeignKeyComparisonWarning(
-          name: 'reference table',
-          expected: referenceTable,
-          found: other.referenceTable,
-        ),
-      );
-    }
-
-    if (referenceTableSchema != other.referenceTableSchema) {
-      mismatches.add(
-        ForeignKeyComparisonWarning(
-          name: 'reference schema',
-          expected: referenceTableSchema,
-          found: other.referenceTableSchema,
-        ),
-      );
-    }
-
-    if (!const ListEquality()
-        .equals(referenceColumns, other.referenceColumns)) {
-      mismatches.add(
-        ForeignKeyComparisonWarning(
-          name: 'reference columns',
-          expected: '$referenceColumns',
-          found: '${other.referenceColumns}',
-        ),
-      );
-    }
-
-    if (onUpdate != other.onUpdate) {
-      mismatches.add(
-        ForeignKeyComparisonWarning(
-          name: 'onUpdate action',
-          expected: '$onUpdate',
-          found: '${other.onUpdate}',
-        ),
-      );
-    }
-
-    if (onDelete != other.onDelete) {
-      mismatches.add(
-        ForeignKeyComparisonWarning(
-          name: 'onDelete action',
-          expected: '$onDelete',
-          found: '${other.onDelete}',
-        ),
-      );
-    }
-
-    if (matchType != null &&
-        other.matchType != null &&
-        matchType != other.matchType) {
-      mismatches.add(
-        ForeignKeyComparisonWarning(
-          name: 'match type',
-          expected: '$matchType',
-          found: '${other.matchType}',
-        ),
-      );
-    }
-
-    return mismatches;
-  }
-}
-
 /// SQL code generation methods for [Filter].
 extension FilterGenerator on Filter {
   /// Generates a SQL WHERE clause from the filter.
@@ -527,22 +259,4 @@ extension FilterConstraintGenerator on FilterConstraint {
 String _microsecondsToInterval(int microseconds) {
   var duration = Duration(microseconds: microseconds);
   return 'INTERVAL \'${duration.inDays} days ${duration.inHours.remainder(24)} hours ${duration.inMinutes.remainder(60)} minutes ${duration.inSeconds.remainder(60)} seconds\'';
-}
-
-extension _ColumnTypeComparison on ColumnType {
-  bool like(ColumnType other) {
-    const integerEquivalentTypes = {
-      ColumnType.integer,
-      ColumnType.bigint,
-    };
-
-    // If this type is in the integerEquivalentTypes set,
-    // check if the other type is also in it.
-    if (integerEquivalentTypes.contains(this)) {
-      return integerEquivalentTypes.contains(other);
-    }
-
-    // Direct comparison for other types.
-    return this == other;
-  }
 }
